HDDS-15394. Moved SnapshotDiffValueParser to om.snapshot.diff package.
